- Add Discount issue fixed - Remove discount button added if added discount - Manual item not showing in list issue fixed - Create order with manual(keypad) item - CashLog UI update - Cash IN/OUT API reason add - Cash Out API issue fixed - cash drawer amount, cash out amount,cash in amount display

+++ app/src/main/java/com/android/pos/data/model/responseModel/CashLogResponse.kt
@@ -17,11 +17,11 @@ data class CashLogResponse(
         @SerializedName("cashes")
         val cashes: List<Cashe>,
         @SerializedName("total_cash_in")
-        val totalCashIn: Int,
+        val totalCashIn: Double,
         @SerializedName("total_cash_out")
-        val totalCashOut: Int,
+        val totalCashOut: Double,
         @SerializedName("total_left_amount_in_drawer")
-        val totalLeftAmountInDrawer: Int
+        val totalLeftAmountInDrawer: Double
     ) {
         data class Cashe(
             @SerializedName("amount")
